requirements.txt
streamlit
pandas
groq
PyMuPDF
langchain-groq
langchain-core
python-dotenv
apscheduler
langchain
beautifulsoup4
requests

